+++ runbot/models/host.py
@@ -1,8 +1,6 @@
 import logging
-import os
-
 from odoo import models, fields, api
-from ..common import fqdn, local_pgadmin_cursor
+from ..common import fqdn, local_pgadmin_cursor, os
 from ..container import docker_build
 _logger = logging.getLogger(__name__)
